Main Menu Current Version
Doesn't work unless you start it from the main menu scene, otherwise it goes straight to the game.  Also Struggling to get the music to carry over from the menu into the actual game.  Just thought I would store it to my branch in case anyone wants to goof with it.